% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCauchy.R
\name{pCauchy}
\alias{pCauchy}
\title{The Cauchy Merging Function}
\usage{
pCauchy(p, method = "A", epi = 0.1, censor = TRUE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{method}{method used to merge p-values, "A" or "I"; by default, method = "A".}

\item{epi}{The significance level of the hypothesis test; takes value in (0,0.5); by default epi = 0.1.}

\item{censor}{logical; if TRUE, p-values are left-censored at 0 and right-censored at 1; by default, censor = TRUE.}
}
\value{
The function returns the details of the hypothesis test including the method used, test result (i.e., reject or fail to reject), the significance level, the test statistic and the threshold.
}
\description{
A function to merge arbitrarily dependent or independent p-values via the Cauchy combination method.
}
\details{
The Cauchy combination method \insertCite{L}{pmerge} calculates the test statistic via
\deqn{ M_{ \mathcal C,K}(p_{1},\dots,p_{K}):=\mathcal C\left(\frac{1}{K}\sum_{k=1}^{K}\mathcal C^{-1}\left(p_{k}\right)\right).}
Here \eqn{\mathcal C} is the cumulative distribution function of the standard Cauchy random variable and \eqn{\mathcal C^{-1}} is the inverse of \eqn{\mathcal C}.
The hypothesis test is rejected if the test statistic is less than or equal to a threshold. The threshold is different for different methods [dependence="I" or "A"].

"I" dependence [dependence="I"]: The threshold of the Cauchy combination method for independent p-values is (asymptotically) equivalent to the significance level; see \insertCite{L;textual}{pmerge}.

"A" dependence [dependence="A"]: See \insertCite{C}{pmerge} for details about the threshold of the Cauchy combination method for arbitrarily dependent p-values.
}
\references{
\insertRef{C}{pmerge}

\insertRef{L}{pmerge}
}
