% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergep.R
\name{phommel}
\alias{phommel}
\title{The Hommel Merging Function.}
\usage{
phommel(p, method = "H", subset = FALSE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{method}{method to merge p-values, "H", "S" or "G"; by default, method="H".}

\item{subset}{logical; if TRUE, p-values less than 0 or greater than 1 are removed; by default, subset=FALSE.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge dependent p-values via the Hommel method and the grid harmonic merging method,  or independent p-values via the Simes method.
}
\details{
Both "H" and "G" methods are valid for dependent p-values. The "S" method is valid for independent p-values.

"H" method [method="H"]: The Hommel method \insertCite{H}{pmerge} calculates the merged p-value by
\deqn{H_{K}(p_{1},\dots,p_{K})=l_{K}\wedge_{k=1}^{K}G_{k,K}(p_{1},\dots,p_{k}),}
where \eqn{p_{(k)}} is the \eqn{k}-th order statistic of the p-values, \eqn{l_{K}=\sum_{k=1}^{K}k^{-1}} and
\deqn{G_{k,K}(p_{1},\dots,p_{K})=\frac{K}{k}p_{(k)}\wedge 1.}

"S" method [method="S"]: The Simes method \insertCite{S}{pmerge} calculates the merged p-value by
\deqn{S_{K}(p_{1},\dots,p_{K})=\frac{1}{l_{K}}H_{K}(p_{1},\dots,p_{K}),}
where \eqn{H_{K}} is given in the details of "H" method.

"G" method [method="G"]: The grid harmonic method in \insertCite{VWW;textual}{pmerge}, which is a modification of the Hommel method, produces a smaller merged p-value than the Hommel method; see \insertCite{VWW;textual}{pmerge} for more details.

A warning is given if any of the p-values is less than zero or greater than one.
}
\note{
Although the Simes method is developed for independent p-values, it is also valid for a wide range of dependency on p-values, e.g., MTP\eqn{_{2}} condition \insertCite{Sarkar}{pmerge}. However, it is not always valid for dependent p-values.
}
\references{
\insertRef{H}{pmerge}

\insertRef{Sarkar}{pmerge}

\insertRef{S}{pmerge}

\insertRef{VWW}{pmerge}
}
