% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergep.R
\name{pmean}
\alias{pmean}
\title{The Generalized Mean Merging Function}
\usage{
pmean(p, r = 0, dependence = "A", subset = FALSE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{r}{the exponent of the generalized mean function in the set \eqn{[-10,\infty]\cup\{-\infty\}};}

\item{dependence}{the dependence assumption of p-values, "A" or "I".}

\item{subset}{logical; if TRUE, p-values less than 0 or greater than 1 are removed; by default, subset=FALSE.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge dependent p-values or independent p-values via the generalized mean method.
}
\details{
The generalized mean merging method combines p-values through the generalized mean function
\deqn{M_{r,K}(p_{1},\dots,p_{K})=\left(\frac{1}{K}\sum_{k=1}^{K}p_{k}^{r}\right)^{\frac{1}{r}},}
including the limiting cases:
\deqn{M_{-\infty,K}(p_{1},\dots,p_{K})=\min\{p_{1},\dots,p_{K}\};}
\deqn{M_{0,K}(p_{1},\dots,p_{K})=\left(\prod_{k=1}^{K}p_{k}\right)^{\frac{1}{K}};}
\deqn{M_{\infty,K}(p_{1},\dots,p_{K})=\max\{p_{1},\dots,p_{K}\}.}

Some special cases of the generalized mean merging method are:

 (1) The Bonferroni method [\eqn{r=-\infty}, method="A"];


 (2) The harmonic averaging method [\eqn{r=-1}, method="A"] \insertCite{VW}{pmerge};

 (3) The Fisher method [\eqn{r=0} and method="I"].

"A" dependence [dependence="A"]: The method is valid for dependent p-values. The merged p-value is calculated by
\deqn{F_{r,K}(p_{1},\dots,p_{K})=b_{r,K}M_{r,K}(p_{1},\dots,p_{K}),}
where \eqn{M_{r,K}} is the generalized mean of the p-values with exponent \eqn{r}, and \eqn{b_{r,K}} is a constant; the constant \eqn{b_{r,K}} can be found in \insertCite{VW;textual}{pmerge} and \insertCite{VWW;textual}{pmerge}.

"I" dependence [dependence="I"]: The method is valid for independent p-values. The merged p-value is approximated by the classic central limit theorem or the generalized central limit theorem; see \insertCite{C;textual}{pmerge} for more details. For exponent less than \eqn{0}, large number of p-values is needed to ensure the accuracy of approximation.

When the number of p-values is large, if \eqn{r\in[-2,1]} and method="A", \eqn{r} is rounded to the closest integer to prevent computational issues and convergence issues; see \insertCite{VW;textual}{pmerge} and \insertCite{VWW;textual}{pmerge}.

Zero p-values are omitted if r is less than zero

A warning is given if any of the p-values is less than zero or greater than one.
}
\references{
\insertRef{C}{pmerge}

\insertRef{VW}{pmerge}

\insertRef{VWW}{pmerge}
}
