% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergep.R
\name{porder}
\alias{porder}
\title{The Order Statistics Merging Function}
\usage{
porder(p, k, censor = TRUE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{k}{the \eqn{k}-th order statistic, takes value \eqn{1,2,\dots,K} where \eqn{K} is the number of p-values.}

\item{censor}{logical; if TRUE, p-values are left-censored at 0 and right-censored at 1; by default, censor = TRUE.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge arbitrarily dependent p-values via order statistics merging method.
}
\details{
The order statistics merging method combines p-values using order statistics and it includes the famous Bonferroni method [k=1, method="O1"].

It is first proposed by \insertCite{R;textual}{pmerge} to merge arbitrarily dependent p-values by
\deqn{G_{k,K}(p_{1},\dots,p_{K})=\frac{K}{k}p_{(k)}\wedge 1,}
where \eqn{p_{(k)}} is the \eqn{k}-th order statistic of the p-values.

This function uses an improved version \insertCite{VWW}{pmerge} of the above order statistics merging method. The improved method returns a smaller merged p-value than the order statistics merging method \insertCite{R}{pmerge}. The merged p-value is calculated by
\deqn{G^{*}_{k,K}(p_{1},\dots,p_{K})=G_{k,K}(p_{1},\dots,p_{K})\wedge 1_{\{p_{(1)}>0\}}.}
where \eqn{p_{(1)}} is the first order statistic of the p-values.

A warning is given if any of the p-values is less than zero or greater than one.
}
\references{
\insertRef{R}{pmerge}

\insertRef{VWW}{pmerge}
}
