% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSimes.R
\name{pSimes}
\alias{pSimes}
\title{The Simes Merging Function.}
\usage{
pSimes(p, method = "H", censor = TRUE, lower = 10^(-10))
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{method}{method to merge p-values, "H", "S" or "G"; by default, method = "H".}

\item{censor}{logical; if TRUE, p-values are left-censored at 0 and right-censored at 1; by default, censor = TRUE.}

\item{lower}{the lower bound for \code{\link{uniroot}} function; by default lower = 1e-10.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge arbitrarily dependent p-values via the Hommel method and the grid harmonic merging method,  or independent p-values via the Simes method.
}
\details{
Both "H" and "G" methods are valid for arbitrarily dependent p-values. The "S" method is valid for independent p-values.

"H" method [method="H"]: The Hommel method \insertCite{H}{pmerge} calculates the merged p-value by
\deqn{H_{K}(p_{1},\dots,p_{K})=\ell_{K}\wedge_{k=1}^{K}G_{k,K}(p_{1},\dots,p_{k}),}
where \eqn{p_{(k)}} is the \eqn{k}-th order statistic of the p-values, \eqn{\ell_{K}=\sum_{k=1}^{K}k^{-1}} and
\deqn{G_{k,K}(p_{1},\dots,p_{K})=\frac{K}{k}p_{(k)}\wedge 1.}

"S" method [method="S"]: The Simes method \insertCite{S}{pmerge} calculates the merged p-value by
\deqn{S_{K}(p_{1},\dots,p_{K})=\frac{1}{\ell_{K}}H_{K}(p_{1},\dots,p_{K}),}
where \eqn{H_{K}} is given in the details of "H" method.

"G" method [method="G"]: The grid harmonic method in \insertCite{VWW;textual}{pmerge}, which is an improvement of the Hommel method, produces a smaller merged p-value than the Hommel method; see \insertCite{VWW;textual}{pmerge} for more details.

A warning is given if any of the p-values is less than zero or greater than one.
}
\note{
Although the Simes method is developed for independent p-values, it is also valid for a wide range of dependency on p-values, e.g., MTP\eqn{_{2}} condition \insertCite{Sarkar}{pmerge}. However, it is not always valid for arbitrarily dependent p-values.
}
\references{
\insertRef{H}{pmerge}

\insertRef{Sarkar}{pmerge}

\insertRef{S}{pmerge}

\insertRef{VWW}{pmerge}
}
