% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergep.R
\name{pharmonic}
\alias{pharmonic}
\title{The Harmonic Mean Merging Function}
\usage{
pharmonic(p, method = "H1", subset = FALSE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{method}{method to merge p-values, "H1" or "H2"; by default, method="H1".}

\item{subset}{logical; if TRUE, p-values less than 0 or greater than 1 are removed; by default, subset=FALSE.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge dependent p-values via the harmonic and the harmonic* merging methods.
}
\details{
Both "H1" and "H2" methods are valid for dependent p-values and they require the number of p-values to be greater than 2.

"H1" method [method="H1"]: The harmonic mean method \insertCite{VW}{pmerge} calculates the merged p-value by
\deqn{F_{-1,K}(p_{1},\dots,p_{K})=b_{-1,K}M_{-1,K}(p_{1},\dots,p_{K}),}
where \eqn{M_{-1,K}(p_{1},\dots,p_{K})=\left(\frac{1}{K}\sum_{k=1}^{K}p_{k}^{-1}\right)^{-1}} is the harmonic mean of the p-values and \eqn{b_{-1,K}} is a constant; see \insertCite{VW;textual}{pmerge} for the calculation of \eqn{b_{-1,K}}.

"H2" method [method="H2"]: The harmonic\eqn{^{*}} merging method \insertCite{VWW}{pmerge}, which is a modification of the harmonic averaging method \insertCite{VW}{pmerge}, returns a smaller merged p-value than the harmonic averaging method; see \insertCite{VWW;textual}{pmerge} for more details.

A warning is given if any of the p-values is less than zero or greater than one.
}
\references{
\insertRef{VW}{pmerge}

\insertRef{VWW}{pmerge}
}
