% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergep.R
\name{pmean}
\alias{pmean}
\title{The Generalized Mean Merging Function}
\usage{
pmean(p, r = 0, dependence = "A", censor = TRUE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{r}{the exponent of the generalized mean function in the set \eqn{[-10,\infty]\cup\{-\infty\}}; by default, r = 0.}

\item{dependence}{the dependence assumption of p-values, "A" or "I"; by default, dependence = "A".}

\item{censor}{logical; if TRUE, p-values are left-censored at 0 and right-censored at 1; by default, censor = TRUE.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge arbitrarily dependent p-values or independent p-values via the generalized mean method.
}
\details{
The generalized mean merging method combines p-values through the generalized mean function
\deqn{M_{r,K}(p_{1},\dots,p_{K})=\left(\frac{1}{K}\sum_{k=1}^{K}p_{k}^{r}\right)^{\frac{1}{r}},}
including the limiting cases:
\deqn{M_{-\infty,K}(p_{1},\dots,p_{K})=\min\{p_{1},\dots,p_{K}\};}
\deqn{M_{0,K}(p_{1},\dots,p_{K})=\left(\prod_{k=1}^{K}p_{k}\right)^{\frac{1}{K}};}
\deqn{M_{\infty,K}(p_{1},\dots,p_{K})=\max\{p_{1},\dots,p_{K}\}.}

Some special cases of the generalized mean merging method are:
\enumerate{
\item
The Bonferroni method for arbitrarily dependent p-values [\eqn{r=-\infty}, method="A"];
\item
The Fisher method for independent p-values [\eqn{r=0} and method="I"];
\item
The harmonic mean method for independent p-values [\eqn{r=-1}, method="I"] \insertCite{W}{pmerge};
\item
The harmonic mean method for arbitrarily dependent p-values [\eqn{r=-1}, method="A"] \insertCite{VW}{pmerge}.
}

"A" dependence [dependence="A"]: The method is valid for dependent p-values. The merged p-value is calculated by
\deqn{F_{r,K}(p_{1},\dots,p_{K})=b_{r,K}M_{r,K}(p_{1},\dots,p_{K}),}
where \eqn{M_{r,K}} is the generalized mean of the p-values with exponent \eqn{r}, and \eqn{b_{r,K}} is a constant; the constant \eqn{b_{r,K}} can be found in \insertCite{VW;textual}{pmerge} and \insertCite{VWW;textual}{pmerge}.

"I" dependence [dependence="I"]: The method is valid for independent p-values. The merged p-value is approximated by the classic central limit theorem or the generalized central limit theorem; see \insertCite{C;textual}{pmerge} for more details.  Also see \strong{Note} for the use of this method.

If \eqn{r\le 0}, any zero p-value will produce a zero merged p-value.

A warning is given if any of the p-values is less than zero or greater than one.
}
\note{
The method for independent p-values [dependence="I"]: The (generalized) central limit theorem has generally good approximations for merging independent p-values except for \eqn{r} close to \eqn{-0.5}. Therefore, we do not recommend to use \eqn{r\in(-1,0)} for independent p-values. For other choices of \eqn{r}, the number of p-values is suggested to be greater than \eqn{100}.
}
\references{
\insertRef{C}{pmerge}

\insertRef{VW}{pmerge}

\insertRef{VWW}{pmerge}

\insertRef{W}{pmerge}
}
