% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergep.R
\name{porder}
\alias{porder}
\title{The Order Statistics Merging Function}
\usage{
porder(p, k, method = "O1", subset = FALSE)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{k}{the \eqn{k}-th order statistic, takes value \eqn{1,2,\dots,K} where \eqn{K} is the number of p-values.}

\item{method}{method used to merge p-values, "O1" or "O2"; by default, method="O1".}

\item{subset}{logical; if TRUE, p-values less than 0 or greater than 1 are removed; by default, subset=FALSE.}
}
\value{
The function returns the merged p-value.
}
\description{
A function to merge dependent p-values via order statistics merging method.
}
\details{
The order statistics merging method combines p-values using order statistics and it includes the famous Bonferroni method [k=1, method="O1"].

Both "O1" and "O2" methods are valid for dependent p-values.

"O1" method [method="O1"]: The order statistics class of merging methods proposed by \insertCite{R;textual}{pmerge} calculates the merged p-value by
\deqn{G_{k,K}(p_{1},\dots,p_{K})=\frac{K}{k}p_{(k)}\wedge 1,}
where \eqn{p_{(k)}} is the \eqn{k}-th order statistic of the p-values.

"O2" method [method="O2"]: It is a modification \insertCite{VWW}{pmerge} of the order statistics merging method \insertCite{R}{pmerge} by a zero-one adjustment. It returns a smaller merged p-value than the order statistics merging method \insertCite{R}{pmerge}. The merged p-value is calculated by
\deqn{G^{*}_{k,K}(p_{1},\dots,p_{K})=G_{k,K}(p_{1},\dots,p_{K})\wedge 1_{\{p_{(1)}>0\}}.}
where \eqn{p_{(1)}} is the first order statistic of the p-values.

A warning is given if any of the p-values is less than zero or greater than one.
}
\references{
\insertRef{R}{pmerge}

\insertRef{VWW}{pmerge}
}
